+++ hosts/nixos/common-nixos.nix
@@ -19,6 +19,7 @@ in
     ../../modules/nixos/home-manager.nix
     ../../modules/nixos/certbot.nix
     ../../modules/nixos/ghostty.nix
+    ../../modules/nixos/agenix-fix.nix
     ../../modules/common/packages.nix
     ../../modules/common/age-host-key.nix
     ../../modules/common/secrets.nix